Unpack and route multiple-element loads to BWQs
This adds preliminary support in the VLU for loads containing multiple
contiguous elements.

At the head of the VLDQ, the subword elements are extracted from the
load data and, for floating-point operations, passed through parallel
recoders.  A permutation network then collectively routes the elements
to their respective BWQs.  As the elements are arranged by consecutive
indices in memory, the permutation simply involves a rotation performed
by circular barrel shifters, with a displacement determined by the utidx
and offset metadata.

The valid signals for individual elements, derived from the utcnt
metadata, are rotated identically to match the BWQs.  The BWQ ready
signals are rotated in the reverse direction, and the utcnt selects
the appropriate subset to reduce for the VLDQ.