Coalesce unit-stride vector loads
For vector loads of unit stride, the VMU now utilizes the full width of
the memory interface to load multiple elements per memory request, as
maximally permitted by the data type size.

In the case of suboptimal base alignment or irregular vector lengths,
at worst only the first and last accesses to a vector are partially
filled; the element count and offset metadata subsequently instruct the
VLU to ignore irrelevant portions of the response.