Fix valve counter decrement
Eliminate the use of io.mbox.bits.meta.ecnt as the subtrahend for the
valve counter.  The width of this signal is deliberately constrained to
tlByteAddrBits so as to reduce the saved metadata state; consequently,
the value tlDataBytes cannot be represented directly but is encoded as 0
instead.  Improper application in arithmetic context caused the valve
counter not to be decremented for vector byte memory operations.