Prevent VLU from prematurely accepting responses
Mark memory responses with an explicit `last' flag so that the VLU can
recognize when to stop accepting further responses until it has
progressed to the next operation.

Note that there could be a brief duration when all elements involved
with the current op have already been received by the VLU, but some
remainder of which has not yet been formally retired.  Without such an
interlock during this interval, the VLU may inadvertently admit loads
intended for the subsequent op and treat them as part of the current op,
leading to incorrect writeback action and deadlock due to missed commit.

Addendum:
To further reduce latency between back-to-back memory operations,
decouple control of the memory interface arbiter from VMU op issue.
It is no longer necessary to wait for all ongoing transactions to
conclude before initiating the next VMU op, which enables address
generation to proceed earlier.