confprec: Optimize expander check
Eliminate false positives in which an older operation is not sequenced
when a younger dependent operation is in the expander, despite the
absence of an actual data hazard since the element ranges are disjoint.

Note that the original check essentially consisted of one side of an
interval comparison -- although correct, it proved over-conservative
since the expander check is agnostic to program order (as it merely
needs to detect overlapping strips).