Enforce loads waiting for store acquires finishing
This fixes a subtle memory ordering bug exposed by
lower latency cache and high lane counts in torture.
We record which sequencer slots are still sending
acquires out and disallow vcu ops issuing before
all previous slots are done sending acquires.

This appears to have a negiligble effect on performance
and is needed for correctness.
We actually only need to enforce this delay if the addresses
overlap but thats much more expensive to check.