vmu: Fix VCU count when final set of predicates is uniformly false
For non-unit strided memory operations, ecnt should be capped by the
remaining vlen if less than the maximum strip size.  Since PBox0 scans
for the longest power-of-2 prefix of false predictes, ecnt_n presents an
inflated count when the last predicate group consists of all zeroes and
the vlen is not a multiple of the strip size, as the unused trailing
bits are also blindly counted.

This caused the VCU counter to be erroneously incremented beyond the
actual vlen and thus fail to return to zero between memory operations.