lane: Fix predicated vcmp* operations
The PRF writeback logic had an incomplete implementation of predication.
Elements that were predicated off for the vcmp* operation were not
correctly masked off when updating the destination predicate register,
as it was only examining io.op.pred.write.bits.pred from LaneCtrl, which
marks all elements in the active strip but does not ignore the disabled
elements among them.

These changes propagate the actual predicates from the lane functional
units to the write mask generation in the PRF.