vmu: Hold element count in vector store table
The element count is now held in the VST instead of being directly
encoded within the TileLink source identifier, which utilizes the
source ID range more efficiently so as to increase the number of
concurrent in-flight stores.  In addition, the size of VST has been
decoupled from that of the VLT.

This may also fix a latent bug with potential source ID collisions
between loads and stores.