vlu: Assert map.io.free only when active
Depending on initial values of non-reset registers, it was possible for
`vlen_end' to be asserted even while the VLU was idle, which drove the
VLU mapper into an inconsistent state (a double-free bug by analogy).

Also ensure that the `used' flag is set in the mapper if a new entry is
allocated during the same cycle that a previous entry is freed.